Changes to user lookups, fix to a module and add command
Level up events are now up to each server's choosing.
You can pick which level gets a level up role.
Additionally I made the level up event a bit more clear.

using the add command got a revamp, new syntaxes and command will be disclosed later if you are not in the test server, it is now possible to add points to everyone with a certain role.